requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
python-jose
passlib[bcrypt]
python-multipart
celery
sqlalchemy
psycopg2-binary
alembic
asyncpg
aiofiles
python-multipart